Magic System Updates

+++ SlimeEnemy.cs
@@ -14,7 +14,7 @@ public class SlimeEnemy : KinematicBody2D
     private Vector2 Velocity;
     private const float Gravity = 200.0f;
     private const float Speed = 30.0f;
-    private int health;
+    private float health;
     
     // Called when the node enters the scene tree for the first time.
     public override void _Ready()
@@ -77,7 +77,7 @@ public class SlimeEnemy : KinematicBody2D
         }
     }
 
-    public void TakeDamage(int damageAmount)
+    public void TakeDamage(float damageAmount)
     {
         health -= damageAmount;
         if(health <= 0)
